# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StudyPilot API",
    "summary": "Phase 1 backend core loop for StudyPilot",
    "version": "0.1.0"
  },
  "paths": {
    "/api/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Upsert Profile",
        "operationId": "upsert_profile_api_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/assets": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Assets",
        "operationId": "list_assets_api_knowledge_assets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeAssetListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload Assets",
        "operationId": "upload_assets_api_knowledge_assets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_assets_api_knowledge_assets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeAssetListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/assets/{asset_id}": {
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Delete Asset",
        "operationId": "delete_asset_api_knowledge_assets__asset_id__delete",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/assets/{asset_id}/retry": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Retry Asset",
        "operationId": "retry_asset_api_knowledge_assets__asset_id__retry_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeAsset"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/search": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search Assets",
        "operationId": "search_assets_api_knowledge_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "topK",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Topk"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workflow/advance": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Advance Workflow",
        "operationId": "advance_workflow_api_workflow_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowAdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowState"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflow/proceed": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Proceed Workflow",
        "operationId": "proceed_workflow_api_workflow_proceed_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowState"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_assets_api_knowledge_assets_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_assets_api_knowledge_assets_post"
      },
      "ChunkSearchResult": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "chunk_id",
          "content",
          "score"
        ],
        "title": "ChunkSearchResult"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeAsset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgeAssetStatus"
          },
          "parse_error_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParseErrorReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          },
          "raw_path": {
            "type": "string",
            "title": "Raw Path"
          },
          "parsed_text": {
            "type": "string",
            "title": "Parsed Text",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeChunk"
            },
            "type": "array",
            "title": "Chunks"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source_type",
          "file_type",
          "file_size_bytes",
          "status",
          "raw_path",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeAsset"
      },
      "KnowledgeAssetListResponse": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeAsset"
            },
            "type": "array",
            "title": "Assets"
          }
        },
        "type": "object",
        "required": [
          "assets"
        ],
        "title": "KnowledgeAssetListResponse"
      },
      "KnowledgeAssetStatus": {
        "type": "string",
        "enum": [
          "uploading",
          "parsing",
          "indexing",
          "ready",
          "parse_failed"
        ],
        "title": "KnowledgeAssetStatus"
      },
      "KnowledgeChunk": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "embedding_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "KnowledgeChunk"
      },
      "KnowledgeSearchResponse": {
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkSearchResult"
            },
            "type": "array",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "chunks"
        ],
        "title": "KnowledgeSearchResponse"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "ParseErrorReason": {
        "type": "string",
        "enum": [
          "unsupported_format",
          "file_corrupted",
          "empty_content"
        ],
        "title": "ParseErrorReason"
      },
...
      "UserProfile": {
        "properties": {
          "learning_goals": {
            "type": "string",
            "title": "Learning Goals"
          },
          "subject_scope": {
            "type": "string",
            "title": "Subject Scope"
          },
          "current_level": {
            "type": "string",
            "title": "Current Level"
          },
          "time_budget": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "object",
            "title": "Time Budget"
          },
          "preferred_style": {
            "type": "string",
            "title": "Preferred Style"
          },
          "preferred_difficulty": {
            "type": "string",
            "title": "Preferred Difficulty"
          },
          "preferred_question_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Question Types"
          },
          "behavior_summary": {
            "type": "string",
            "title": "Behavior Summary",
            "default": ""
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "learning_goals",
          "subject_scope",
          "current_level",
          "time_budget",
          "preferred_style",
          "preferred_difficulty",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "UserProfile"
      },
      "UserProfileCreate": {
        "properties": {
          "learning_goals": {
            "type": "string",
            "title": "Learning Goals"
          },
          "subject_scope": {
            "type": "string",
            "title": "Subject Scope"
          },
          "current_level": {
            "type": "string",
            "title": "Current Level"
          },
          "time_budget": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "object",
            "title": "Time Budget"
          },
          "preferred_style": {
            "type": "string",
            "title": "Preferred Style"
          },
          "preferred_difficulty": {
            "type": "string",
            "title": "Preferred Difficulty"
          },
          "preferred_question_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Question Types"
          },
          "behavior_summary": {
            "type": "string",
            "title": "Behavior Summary",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "learning_goals",
          "subject_scope",
          "current_level",
          "time_budget",
          "preferred_style",
          "preferred_difficulty"
        ],
        "title": "UserProfileCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowAdvanceRequest": {
        "properties": {
          "next_stage": {
            "$ref": "#/components/schemas/WorkflowStage"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "next_stage",
          "reason"
        ],
        "title": "WorkflowAdvanceRequest"
      },
      "WorkflowStage": {
        "type": "string",
        "enum": [
          "onboarding",
          "planning",
          "learning",
          "checkpoint_test",
          "unit_review",
          "unit_test",
          "final_review",
          "final_test",
          "next_cycle"
        ],
        "title": "WorkflowStage"
      },
      "WorkflowState": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "current_stage": {
            "$ref": "#/components/schemas/WorkflowStage"
          },
          "current_macro_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Macro Plan Id"
          },
          "current_micro_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Micro Plan Id"
          },
          "recent_assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Assessment Id"
          },
          "next_action": {
            "type": "string",
            "title": "Next Action"
          },
          "stage_history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Stage History"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "current_stage",
          "next_action",
          "updated_at"
        ],
        "title": "WorkflowState"
      }
    }
  }
}